move PCPManager to cp_utils.py
